project_requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==1.11.24
django-allauth==0.40.0
django-extra-views==0.12.0
djangorestframework==3.9.4
idna==2.8
oauthlib==3.1.0
Pillow==6.1.0
PyMySQL==0.9.3
python-openid==2.2.5
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
sentry-sdk==0.11.2
six==1.12.0
urllib3==1.25.3
